코드 리펙토링 enum별도 코드파일 및 다른 코드들 가독성 좋게 코드 분할
