+++ src/main/java/lotto/LottoController/CalculatorResult.java
@@ -1,6 +1,7 @@
 package lotto.LottoController;
 
-import lotto.Lotto;
+import lotto.LottoModel.Lotto;
+import lotto.LottoModel.LottoRank;
 
 import java.text.NumberFormat;
 import java.util.HashMap;
@@ -10,29 +11,43 @@ import java.util.Map;
 
 public class CalculatorResult {
 
-    public Map<Lotto.LottoRank, Integer> calculateResults(List<Lotto> lottos, Lotto winNumbers, int bonusNumber) {
-        Map<Lotto.LottoRank, Integer> result = new HashMap<>();
+    public Map<LottoRank, Integer> calculateResults(List<Lotto> lottos, Lotto winNumbers, int bonusNumber) {
+        Map<LottoRank, Integer> result = initializeResult();
 
-        for (Lotto.LottoRank rank : Lotto.LottoRank.values()) {
-            result.put(rank, 0);
+        for (Lotto lotto : lottos) {
+            LottoRank rank = lotto.determineRank(winNumbers, bonusNumber);
+            result.put(rank, result.get(rank) + 1);
         }
 
-        for (Lotto currntLotto : lottos) {
-            Lotto.LottoRank rank = currntLotto.determineRank(winNumbers, bonusNumber);
-            result.put(rank, result.get(rank) + 1);
+        return result;
+    }
+
+    private Map<LottoRank, Integer> initializeResult() {
+        Map<LottoRank, Integer> result = new HashMap<>();
+
+        for (LottoRank rank : LottoRank.values()) {
+            result.put(rank, 0);
         }
         return result;
     }
 
-    public String calculateProfit(Map<Lotto.LottoRank, Integer> result, int purchaseAmount) {
-        int totalPrize = 0;
-
-        for (Map.Entry<Lotto.LottoRank, Integer> entry : result.entrySet()) {
-            totalPrize += entry.getKey().getPrize() * entry.getValue();
-        }
-
+    public String calculateProfit(Map<LottoRank, Integer> result, int purchaseAmount) {
+        int totalPrize = TotalPrize(result);
         double profitRate = (double) totalPrize / purchaseAmount * 100;
 
+        return formatProfit(profitRate);
+    }
+
+    private int TotalPrize(Map<LottoRank, Integer> result) {
+        int totalPrize = 0;
+
+        for (Map.Entry<LottoRank, Integer> entry : result.entrySet()) {
+            totalPrize += entry.getKey().getPrize() * entry.getValue();
+        }
+        return totalPrize;
+    }
+
+    private String formatProfit(double profitRate) {
         NumberFormat numberFormat = NumberFormat.getNumberInstance(Locale.getDefault());
 
         return numberFormat.format(profitRate) + "%";
