+++ src/main/java/lotto/LottoController/LottoController.java
@@ -1,10 +1,9 @@
 package lotto.LottoController;
 
-import lotto.Lotto;
+import lotto.LottoModel.Lotto;
+import lotto.LottoModel.LottoRank;
 import lotto.LottoView.LottoView;
-import camp.nextstep.edu.missionutils.Randoms;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -29,20 +28,23 @@ public class LottoController {
         int purchaseAmount = inpubCheck.inputAmount();
         List<Lotto> lottos = generatorLotto.generatorLotto(purchaseAmount);
         view.showLottos(lottos.size(), lottos.stream().map(Lotto::getNumbers).collect(Collectors.toList()));
+
         Lotto winNumbers = inpubCheck.inputNumbers();
         int bonusNumber = inpubCheck.inputBonus(winNumbers);
-        Map<Lotto.LottoRank, Integer> result = calculatorResult.calculateResults(lottos, winNumbers, bonusNumber);
+
+        Map<LottoRank, Integer> result = calculatorResult.calculateResults(lottos, winNumbers, bonusNumber);
+
         showResult(result);
         showProfit(result, purchaseAmount);
     }
 
-    private void showResult(Map<Lotto.LottoRank, Integer> result) {
+    private void showResult(Map<LottoRank, Integer> result) {
         Map<String, Integer> formatedResult = numberFormatter.formatResult(result);
 
         view.showResult(formatedResult);
     }
 
-    private void showProfit(Map<Lotto.LottoRank, Integer> result, int purchaseAmount) {
+    private void showProfit(Map<LottoRank, Integer> result, int purchaseAmount) {
         String profitRate = calculatorResult.calculateProfit(result, purchaseAmount);
 
         view.showRate(profitRate);
