+++ src/main/java/lotto/LottoController/NumberFormatter.java
@@ -1,6 +1,6 @@
 package lotto.LottoController;
 
-import lotto.Lotto;
+import lotto.LottoModel.LottoRank;
 
 import java.text.NumberFormat;
 import java.util.LinkedHashMap;
@@ -10,26 +10,25 @@ import java.util.Map;
 
 public class NumberFormatter {
 
-    public Map<String, Integer> formatResult(Map<Lotto.LottoRank, Integer> result) {
+    public Map<String, Integer> formatResult(Map<LottoRank, Integer> result) {
         Map<String, Integer> formattedResult = new LinkedHashMap<>();
 
-        for (int i = Lotto.LottoRank.values().length - 2; i >= 0; i--) {
-            Lotto.LottoRank rank = Lotto.LottoRank.values()[i];
+        for (int i = LottoRank.values().length - 2; i >= 0; i--) {
+            LottoRank rank = LottoRank.values()[i];
             String showRank = showFormatter(rank);
+
             formattedResult.put(showRank, result.getOrDefault(rank, 0));
         }
         return formattedResult;
     }
 
-    private String showFormatter(Lotto.LottoRank rank) {
+    private String showFormatter(LottoRank rank) {
         NumberFormat numberFormat = NumberFormat.getNumberInstance(Locale.getDefault());
         String prizeFormatted = numberFormat.format(rank.getPrize());
 
-        if (rank == Lotto.LottoRank.SECOND) {
+        if (rank == LottoRank.SECOND) {
             return rank.getMatchCount() + "개 일치, 보너스 볼 일치 (" + prizeFormatted + "원)";
         }
-        else {
-            return rank.getMatchCount() + "개 일치 (" + prizeFormatted + "원)";
-        }
+        return rank.getMatchCount() + "개 일치 (" + prizeFormatted + "원)";
     }
 }
