+++ src/main/java/lotto/LottoModel/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.LottoModel;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -9,7 +9,6 @@ public class Lotto {
     private List<Integer> numbers = new ArrayList<>();
 
     public Lotto(List<Integer> numbers) {
-
         validate(numbers);
         this.numbers = new ArrayList<>(numbers);
         Collections.sort(this.numbers);
@@ -74,40 +73,4 @@ public class Lotto {
         return LottoRank.rankCheck(matchCount, matchbonus);
     }
 
-    public enum LottoRank {
-        FIRST(6, 2_000_000_000, false),
-        SECOND(5, 30_000_000, false),
-        THIRD(5, 1_500_000, false),
-        FOURTH(4, 50_000, false),
-        FIFTH(3, 5_000, false),
-        NONE(0, 0, false);
-
-        private final int matchCount;
-        private final int prize;
-        private final boolean matchBonus;
-
-        LottoRank(int matchCount, int prize, boolean matchBonus) {
-            this.matchCount = matchCount;
-            this.prize = prize;
-            this.matchBonus = matchBonus;
-        }
-
-        public static LottoRank rankCheck(int matchCount, boolean matchBonus) {
-            for (LottoRank rank : values()) {
-
-                if (rank.matchCount == matchCount && (!rank.matchBonus || matchBonus)) {
-                    return rank;
-                }
-            }
-            return NONE;
-        }
-
-        public int getMatchCount() {
-            return matchCount;
-        }
-
-        public int getPrize() {
-            return prize;
-        }
-    }
 }
