+++ src/test/java/lotto/ApplicationTest.java
@@ -55,105 +55,6 @@ class ApplicationTest extends NsTest {
             assertThat(output()).contains(ERROR_MESSAGE);
         });
     }
-    @DisplayName("구매 금액이 1000원 단위가 아닐 때 예외를 발생한다")
-    @Test
-    void 예외_테스트2() {
-        assertSimpleTest(() -> {
-            runException("1500");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("보너스 번호가 당첨번호와 중복되면 예외를 발생한다.")
-    @Test
-    void 예외_테스트3() {
-        assertSimpleTest(() -> {
-            runException("1000", "1,2,3,4,5,6", "3");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("보너스 번호가 1~45범위를 넘어가면 예외를 발생한다.")
-    @Test
-    void 예외_테스트4() {
-        assertSimpleTest(() -> {
-            runException("1000", "1,2,3,4,5,6", "77");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("로또개수 출력")
-    @Test
-    void 예외_테스트5() {
-        assertSimpleTest(() -> {
-            runException("8000", "1,2,3,4,5,6", "77");
-            assertThat(output()).contains("8개를 구매했습니다.");
-        });
-    }
-
-    @DisplayName("당첨번호가 잘나오는지 테스트")
-    @Test
-    void 예외_테스트6() {
-        assertRandomUniqueNumbersInRangeTest(
-                () -> {
-                    run("8000", "1,2,3,4,5,6", "7");
-                    assertThat(output()).contains(
-                            "8개를 구매했습니다.",
-                            "[8, 21, 23, 41, 42, 43]",
-                            "[3, 5, 11, 16, 32, 38]",
-                            "[7, 11, 16, 35, 36, 44]",
-                            "[1, 8, 11, 31, 41, 42]",
-                            "[13, 14, 16, 38, 42, 45]",
-                            "[7, 11, 30, 40, 42, 43]",
-                            "[2, 13, 22, 32, 38, 45]",
-                            "[1, 3, 5, 14, 22, 45]"
-                    );
-                },
-                List.of(8, 21, 23, 41, 42, 43),
-                List.of(3, 5, 11, 16, 32, 38),
-                List.of(7, 11, 16, 35, 36, 44),
-                List.of(1, 8, 11, 31, 41, 42),
-                List.of(13, 14, 16, 38, 42, 45),
-                List.of(7, 11, 30, 40, 42, 43),
-                List.of(2, 13, 22, 32, 38, 45),
-                List.of(1, 3, 5, 14, 22, 45)
-        );
-    }
-
-    @DisplayName("당첨내역 확인 테스트")
-    @Test
-    void 예외_테스트7() {
-        assertRandomUniqueNumbersInRangeTest(
-                () -> {
-                    run("8000", "1,2,3,4,5,6", "7");
-                    assertThat(output()).contains(
-                            "8개를 구매했습니다.",
-                            "[8, 21, 23, 41, 42, 43]",
-                            "[3, 5, 11, 16, 32, 38]",
-                            "[7, 11, 16, 35, 36, 44]",
-                            "[1, 8, 11, 31, 41, 42]",
-                            "[13, 14, 16, 38, 42, 45]",
-                            "[7, 11, 30, 40, 42, 43]",
-                            "[2, 13, 22, 32, 38, 45]",
-                            "[1, 3, 5, 14, 22, 45]",
-                            "3개 일치 (5,000원) - 1개",
-                            "4개 일치 (50,000원) - 0개",
-                            "5개 일치 (1,500,000원) - 0개",
-                            "5개 일치, 보너스 볼 일치 (30,000,000원) - 0개",
-                            "6개 일치 (2,000,000,000원) - 0개"
-                    );
-                },
-                List.of(8, 21, 23, 41, 42, 43),
-                List.of(3, 5, 11, 16, 32, 38),
-                List.of(7, 11, 16, 35, 36, 44),
-                List.of(1, 8, 11, 31, 41, 42),
-                List.of(13, 14, 16, 38, 42, 45),
-                List.of(7, 11, 30, 40, 42, 43),
-                List.of(2, 13, 22, 32, 38, 45),
-                List.of(1, 3, 5, 14, 22, 45)
-        );
-    }
-
 
     @Override
     public void runMain() {
